create components page, install react-select and create select component
