+++ package.json
@@ -18,6 +18,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.51.2",
+    "react-select": "^5.8.0",
     "server-only": "^0.0.1",
     "zod": "^3.22.4",
     "zustand": "^4.5.2"
